Incomprehensible fix for incomprehensible bug
For reasons that are utterly beyond me, it seems that /somehow/ the if
statement in `org-glossary--print-terms-singular' would append the both
the true/false values together when the predicate is non-nil! It is hard
to be certain of this though, because re-evaluating this function (even
when the body was unchanged) changed the behaviour.

For now, reversing the list of extra parameters makes this behave as
expected. Perhaps I could dig deeper, but I fear for my sanity should I
continue.