Properly fix the bizarre issue of 90137ed
I got help from an elisp expert, who explained the black magic behind
the strange behaviour. It seems like the combination of destructive
modification (from nreverse down the line) and a quoted string literal
results in undefined behaviour.

So, unquoting the string literal fixes the problem, but I think the
reverse is a decent idea still. It makes the result more predictable and
the function purer.