feat: support multiple collection roots
Support org-glossary-collection-root to be a list of collection roots.

Each directory will be tested for a match for the short-names in
`#+glossary_sources:`. Multiple files can be found for each `short-name`.

Context/rationale:
I have different contexts such as "public" or "work" where I store
glossaries. "public" glossaries are distributed with my emacs configuration,
"work" glossaries are stored in a dedicated repository.

Considerations:
- error handling: org-glossary reports errors for each short-name that
  cannot be found by returning the collection root concatenated with the
  short-name.

  For multiple collection-roots this will generate an error message for
  each collection-root not containing a short-name