Use the alias term as the value across types
When an alias goes from one category/type of gloss to another, and
producing a non-canonicalised entry, the value should be set to the term
referenced as opposed to the value of the canonical term.

Otherwise the template for one type of gloss can be mixed with the
content of another type of gloss, and that is likely to produce
undesirable output.